[OPT-2][OPT-11] - Excel parsers were added. Decimal was relaced by double. Tests were added for desktop.

+++ OPTEL.Data/CoolingLipChange.cs
@@ -6,9 +6,9 @@ namespace OPTEL.Data
     {
         public int ID { get; set; }
 
-        public decimal CoolingLipToChange { get; set; }
+        public double CoolingLipToChange { get; set; }
 
-        public decimal ReconfigurationTime { get; set; }
+        public double ReconfigurationTime { get; set; }
 
         [Required]
         public virtual ProductionLine ParentProductionLine { get; set; }
